[refactor] stamp, qrscan 패키지 분리
HandleQRScanExcepton 네이밍 변경, HandleException 함수와 명확히 구분을 위함
