+++ feature/stamp/src/main/kotlin/com/unifest/android/feature/stamp/viewmodel/qrscan/QRScanUiAction.kt
@@ -1,4 +1,4 @@
-package com.unifest.android.feature.stamp.viewmodel
+package com.unifest.android.feature.stamp.viewmodel.qrscan
 
 sealed interface QRScanUiAction {
     data class OnQRCodeScanned(val boothId: Long) : QRScanUiAction
