+++ feature/stamp/src/main/kotlin/com/unifest/android/feature/stamp/viewmodel/stamp/StampUiAction.kt
@@ -1,4 +1,4 @@
-package com.unifest.android.feature.stamp.viewmodel
+package com.unifest.android.feature.stamp.viewmodel.stamp
 
 import com.unifest.android.core.common.PermissionDialogButtonType
 import com.unifest.android.core.model.StampFestivalModel
